astro/requirements.txt
scikit-learn==1.6.1
scipy==1.14.1
numpy==1.26.4
pandas==2.2.3
joblib==1.4.2
mlflow==2.18.0
google-cloud-bigquery==3.27.0
google-cloud-bigquery-storage==2.27.0
db-dtypes==1.3.1
pyarrow==18.1.0
python-dotenv==1.0.1
huggingface_hub>=0.27.0
